Broken up FilterToolbar into one more component for easier looping
